Extract Address validation into Address.Validator module
